[advertisement] Tick immediately on start
When we start advertising it takes as long as the heartbeat interval
duration to begin advertising. Ideally a service shouldn't start
advertising until it's ready, so when it starts it might make sense to
show the advertisement immediately. Even if it's not ready at first, if
it has a health function it won't start heartbeating until it's ready
anyway.