[kv] Support watch start revision in watch store
Some use cases require specifying the start revision of a watch request.
Since our only use case so far needs the start revision at a store level
as opposed to on a per-watch basis, this PR implements watch revision
support at the store level.

Includes tests to make sure that we can safely specify a start revision
that has been compacted or one that is in the future with no issues.